+++ src/Requests/PersonRequest.php
@@ -3,6 +3,7 @@
 
 namespace CTApi\Requests;
 
+use CTApi\Exceptions\CTModelException;
 use CTApi\Models\Person;
 
 class PersonRequest
@@ -50,5 +51,19 @@ class PersonRequest
     public static function update(Person $person, array $attributesToUpdate = []): void
     {
         (new PersonRequestBuilder())->update($person, $attributesToUpdate);
+    }
+
+    /**
+     * Delete the person on churchtools.
+     */
+    public static function delete(Person $person): void
+    {
+        $id = $person->getId();
+
+        if (is_null($id)) {
+            throw new CTModelException("ID of Person cannot be null.");
+        }
+
+        (new PersonRequestBuilder())->delete($id);
     }
 }